requirements.txt
autograd==1.4
dill==0.3.5.1
experiments==0.1.0
lxml==4.8.0
matplotlib==3.3.4
networkx==2.8.5
numpy==1.20.1
pandas==1.2.4
psycopg2_binary==2.9.5
pymoo==0.5.0
python-dotenv==0.21.0
scikit_learn==1.1.3
scipy==1.8.0
seaborn==0.11.2
tqdm==4.59.0
